+++ RentACarExt/ICRUD.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace RentACarExt
+{
+    public interface ICRUD
+    {
+        void Create();
+        string ToString(); // Read
+        void Update();
+        void Delete();
+
+
+    }
+}